+++ configs/typescript.cjs
@@ -1,10 +1,14 @@
-/** @type {import('eslint').Linter.Config} */
+/** @type {import("eslint").Linter.Config} */
 module.exports = {
+  extends: "./javascript.cjs",
   overrides: [{
     files: "*.ts",
+    parserOptions: {
+      project: true
+    },
     extends: [
-      "plugin:@typescript-eslint/recommended",
-      "plugin:@typescript-eslint/stylistic"
+      "plugin:@typescript-eslint/strict-type-checked",
+      "plugin:@typescript-eslint/stylistic-type-checked"
     ]
   }]
 };