+++ rulesets/typescript.cjs
@@ -1,7 +1,0 @@
-/** @type {import('eslint').Linter.Config} */
-module.exports = {
-  extends: [
-    "plugin:@typescript-eslint/recommended",
-    "plugin:@typescript-eslint/stylistic"
-  ]
-};